+++ R/ggfoundry-package.R
@@ -5,7 +5,8 @@
 #' @import ggplot2
 #' @import grid
 #' @import grImport2
+#' @importFrom cli cli_abort
 #' @importFrom lifecycle deprecated
-#' @importFrom rlang abort is_empty
+#' @importFrom rlang is_empty
 ## usethis namespace: end
 NULL