+++ tests/testthat/test-geom_casting.R
@@ -1,11 +1,11 @@
 test_that("fills & manual shapes", {
   file <- tempfile(fileext = ".png")
 
-  p <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg, fill = factor(cyl))) +
-    geom_casting(ggplot2::aes(shape = factor(cyl))) +
-    ggplot2::scale_shape_manual(values = c("violin", "box", "dendro"))
+  p <- ggplot(mtcars, aes(wt, mpg, fill = factor(cyl))) +
+    geom_casting(aes(shape = factor(cyl))) +
+    scale_shape_manual(values = c("violin", "box", "dendro"))
 
-  ggplot2::ggsave(file, p, "png")
+  ggsave(file, p, "png")
 
   expect_snapshot_file(file)
 
@@ -14,7 +14,7 @@ test_that("fills & manual shapes", {
 
 test_that("bad shape", {
   expect_snapshot(
-    ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
+    ggplot(mtcars, aes(wt, mpg)) +
       geom_casting(shape = "non-shape"),
     error = TRUE
   )
@@ -29,12 +29,26 @@ test_that("available sets & shapes", {
 test_that("shapes clipped when zooming", {
   file <- tempfile(fileext = ".png")
 
-  p <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) +
+  p <- ggplot(mtcars, aes(wt, mpg)) +
     geom_casting(size = 0.1, shape = "violin") +
-    ggplot2::geom_point() +
-    ggplot2::coord_cartesian(xlim = c(1, 4))
+    geom_point() +
+    coord_cartesian(xlim = c(1, 4))
 
-  ggplot2::ggsave(file, p, "png")
+  ggsave(file, p, "png")
+
+  expect_snapshot_file(file)
+
+  unlink(file)
+})
+
+test_that("Display a palette", {
+  file <- tempfile(fileext = ".png")
+
+  p <- display_palette(
+    c("red", "blue"), 2, "Example", colour = "grey", shape = "jar"
+    )
+
+  ggsave(file, p, "png")
 
   expect_snapshot_file(file)